Created BaseTimelineWidget as a base class for NPSGraphWidget and NoteLayoutWidget
