+++ app/src/gui/widget/NoteLayoutWidget.ts
@@ -1,44 +1,27 @@
-import {
-  FederatedPointerEvent,
-  ParticleContainer,
-  RenderTexture,
-  Sprite,
-  Texture,
-} from "pixi.js"
-import { EditMode } from "../../chart/ChartManager"
+import { ParticleContainer, RenderTexture, Sprite, Texture } from "pixi.js"
 import { QUANT_COLORS } from "../../chart/component/edit/SnapContainer"
-import { Chart } from "../../chart/sm/Chart"
 import { isHoldNote } from "../../chart/sm/NoteTypes"
-import { BetterRoundedRect } from "../../util/BetterRoundedRect"
-import { EventHandler } from "../../util/EventHandler"
-import { Flags } from "../../util/Flags"
-import { clamp, lerp, maxArr, minArr, unlerp } from "../../util/Math"
+import { unlerp } from "../../util/Math"
 import { Options } from "../../util/Options"
-import { destroyChildIf, getDivision, getNoteEnd } from "../../util/Util"
-import { Widget } from "./Widget"
+import { destroyChildIf, getDivision } from "../../util/Util"
+import { BaseTimelineWidget } from "./BaseTimelineWidget"
 import { WidgetManager } from "./WidgetManager"
 
-export class NoteLayoutWidget extends Widget {
+export class NoteLayoutWidget extends BaseTimelineWidget {
   barContainer = new ParticleContainer(
     1500,
     { position: true, scale: true, tint: true },
     16384,
     true
   )
-  backing: BetterRoundedRect = new BetterRoundedRect()
+
   bars: Sprite
   barTexture: RenderTexture
-  overlay: Sprite = new Sprite(Texture.WHITE)
-  selectionOverlay: Sprite = new Sprite(Texture.WHITE)
-
-  private lastHeight = 0
-  private lastCMod
-  private mouseDown = false
-  private queued = false
 
   constructor(manager: WidgetManager) {
     super(manager)
     this.addChild(this.backing)
+    this.addChild(this.container)
     this.visible = false
 
     this.backing.tint = 0
@@ -50,165 +33,8 @@ export class NoteLayoutWidget extends Widget {
 
     this.bars = new Sprite(this.barTexture)
     this.bars.anchor.set(0.5)
-    this.addChild(this.bars)
-
-    this.overlay.anchor.x = 0.5
-    this.overlay.anchor.y = 0
-    this.overlay.alpha = 0.3
-
-    this.selectionOverlay.anchor.x = 0.5
-    this.selectionOverlay.anchor.y = 0
-    this.selectionOverlay.alpha = 0.3
-    this.selectionOverlay.tint = 0x4e6fa3
-
-    this.lastCMod = Options.chart.CMod
-    this.addChild(this.overlay)
-    this.addChild(this.selectionOverlay)
-    this.x = this.manager.app.renderer.screen.width / 2 - 20
-    EventHandler.on("chartLoaded", () => {
-      this.queued = false
-      this.populate()
-    })
-    EventHandler.on("chartModifiedAfter", () => {
-      if (!this.queued) this.populate()
-      this.queued = true
-    })
-    const interval = setInterval(() => {
-      if (this.queued) {
-        this.queued = false
-        this.populate()
-      }
-    }, 3000)
-
-    this.on("destroyed", () => clearInterval(interval))
+    this.container.addChild(this.bars)
     this.populate()
-
-    this.eventMode = "static"
-    this.on("mousedown", event => {
-      this.mouseDown = true
-      this.handleMouse(event)
-    })
-    this.on("mousemove", event => {
-      if (this.mouseDown) this.handleMouse(event)
-    })
-    window.onmouseup = () => {
-      this.mouseDown = false
-    }
-  }
-
-  private handleMouse(event: FederatedPointerEvent) {
-    if (this.manager.chartManager.getMode() == EditMode.Play) return
-    if (!this.getChart()) return
-    let t =
-      (this.bars.toLocal(event.global).y + this.bars.height / 2) /
-      this.bars.height
-    t = clamp(t, 0, 1)
-    const lastNote = this.getChart().getNotedata().at(-1)
-    if (!lastNote) return
-    const lastBeat = lastNote.beat + (isHoldNote(lastNote) ? lastNote.hold : 0)
-    const lastSecond = this.getChart().getSecondsFromBeat(lastBeat)
-    if (Options.chart.CMod) {
-      this.manager.chartManager.setTime(
-        lerp(-this.getChart().timingData.getOffset(), lastSecond, t)
-      )
-    } else {
-      this.manager.chartManager.setBeat(lastBeat * t)
-    }
-  }
-
-  update() {
-    this.scale.y = Options.chart.reverse ? -1 : 1
-    const height = this.manager.app.renderer.screen.height - 40
-    this.backing.height = height + 10
-    this.backing.position.y = -this.backing.height / 2
-    this.backing.position.x = -this.backing.width / 2
-    this.bars.height = height
-    this.x = this.manager.app.renderer.screen.width / 2 - 20
-    const chart = this.getChart()
-    const chartView = this.manager.chartManager.chartView!
-    if (!chart || !chartView || !Flags.layout) {
-      this.visible = false
-      return
-    }
-    this.visible = true
-    const lastNote = chart.getNotedata().at(-1)
-    if (!lastNote) {
-      this.overlay.height = 0
-      return
-    }
-
-    const overlayStart = Options.chart.CMod
-      ? chartView.getSecondFromYPos(
-          -this.manager.app.renderer.screen.height / 2
-        )
-      : chartView.getBeatFromYPos(
-          -this.manager.app.renderer.screen.height / 2,
-          true
-        )
-    const overlayEnd = Options.chart.CMod
-      ? chartView.getSecondFromYPos(this.manager.app.renderer.screen.height / 2)
-      : chartView.getBeatFromYPos(
-          this.manager.app.renderer.screen.height / 2,
-          true
-        )
-    const overlayRange = this.getYFromRange(chart, overlayStart, overlayEnd)
-    this.overlay.y = overlayRange.startY
-    this.overlay.height = overlayRange.endY - overlayRange.startY
-    this.overlay.height = Math.max(2, this.overlay.height)
-
-    const selection = this.manager.chartManager.selection.notes
-    if (selection.length < 1) {
-      this.selectionOverlay.visible = false
-    } else {
-      this.selectionOverlay.visible = true
-      let selectionStart, selectionEnd
-      if (Options.chart.CMod) {
-        selectionStart = minArr(selection.map(note => note.second))
-        selectionEnd = maxArr(
-          selection.map(note => chart.getSecondsFromBeat(getNoteEnd(note)))
-        )
-      } else {
-        selectionStart = minArr(selection.map(note => note.beat))
-        selectionEnd = maxArr(selection.map(note => getNoteEnd(note)))
-      }
-      const selectionRange = this.getYFromRange(
-        chart,
-        selectionStart,
-        selectionEnd
-      )
-      this.selectionOverlay.y = selectionRange.startY
-      this.selectionOverlay.height = selectionRange.endY - selectionRange.startY
-      this.selectionOverlay.height = Math.max(2, this.selectionOverlay.height)
-    }
-
-    if (
-      this.manager.app.renderer.screen.height != this.lastHeight ||
-      this.lastCMod != Options.chart.CMod
-    ) {
-      this.lastCMod = Options.chart.CMod
-      this.lastHeight = this.manager.app.renderer.screen.height
-      this.populate()
-    }
-  }
-
-  private getYFromRange(chart: Chart, start: number, end: number) {
-    const lastBeat = getNoteEnd(chart.getNotedata().at(-1)!)
-    const lastSecond = chart.getSecondsFromBeat(lastBeat)
-    let t_startY = unlerp(0, lastBeat, start)
-    let t_endY = unlerp(0, lastBeat, end)
-    if (Options.chart.CMod) {
-      t_startY = unlerp(-chart.timingData.getOffset(), lastSecond, start)
-      t_endY = unlerp(-chart.timingData.getOffset(), lastSecond, end)
-    }
-    t_startY = clamp(t_startY, 0, 1)
-    t_endY = clamp(t_endY, 0, 1)
-    if (t_startY > t_endY) [t_startY, t_endY] = [t_endY, t_startY]
-    const startY = (t_startY - 0.5) * (this.backing.height - 10)
-    const endY = (t_endY - 0.5) * (this.backing.height - 10)
-    return {
-      startY,
-      endY,
-    }
   }
 
   populate() {
@@ -226,12 +52,6 @@ export class NoteLayoutWidget extends Widget {
     const lastNote = chart.getNotedata().at(-1)
 
     const height = this.manager.app.renderer.screen.height - 40
-    this.backing.height = height
-    this.backing.width = numCols * 6 + 8
-    this.overlay.width = numCols * 6 + 8
-    this.selectionOverlay.width = numCols * 6 + 8
-    this.pivot.x = this.backing.width / 2
-
     this.barTexture.resize(numCols * 6, height)
 
     if (!lastNote) {
@@ -295,9 +115,5 @@ export class NoteLayoutWidget extends Widget {
     this.manager.app.renderer.render(this.barContainer, {
       renderTexture: this.barTexture,
     })
-  }
-
-  private getChart(): Chart {
-    return this.manager.chartManager.loadedChart!
   }
 }
